修改图标

+++ basemodule/src/main/java/com/example/xiaojin20135/basemodule/mpchart/MyValueFormatter.java
@@ -13,6 +13,6 @@ public class MyValueFormatter implements IValueFormatter {
     @Override
     public String getFormattedValue (float value, Entry entry, int dataSetIndex, ViewPortHandler
         viewPortHandler) {
-        return (int)value+"";
+        return value+"";
     }
 }
