+++ basemodule/src/main/java/com/example/xiaojin20135/basemodule/mpchart/fragment/BarChartFrag.java
@@ -156,7 +156,7 @@ public class BarChartFrag extends Fragment {
             if(barChartBaseBean.getColor () != 0){
                 colors.add (barChartBaseBean.getColor ());
             }
-            sizes.add (new BarEntry ((int)(i+1),(int)barChartBaseBean.getCount ()));
+            sizes.add (new BarEntry ((int)(i+1),barChartBaseBean.getCount ()));
         }
         //横向刻度名称
         xAxis.setValueFormatter(new CommonAxisValueFormatter (names));
